feat: 数個の汎用的なメソッドを実装

+++ emia/emulator.hpp
@@ -10,12 +10,12 @@ class Emulator;
 class DirectiveTable {
 
 protected:
-	std::unordered_map<u64, std::function<u64(Emulator *)>> directives;
+	std::unordered_map<u64, std::function<u8(Emulator *)>> directives;
 
-	static u64 mov_r64_imm64(Emulator *emulator);
-	static u64 jmp_rel8(Emulator *emulator);
+	static u8 mov_r64_imm64(Emulator *emulator);
+	static u8 jmp_rel8(Emulator *emulator);
 
-	static u64 add_rm64_imm8(Emulator *emulator);
+	static u8 add_rm64_imm8(Emulator *emulator);
 
 public:
 	DirectiveTable();
